requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.4
django-extensions==4.1
django-redis==6.0.0
djangorestframework==3.17.1
drf-yasg==1.21.15
inflection==0.5.1
kombu==5.6.2
packaging==26.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
